requirements.txt
torch==2.0.0 
xformers
numpy
tyro
diffusers
dearpygui
einops
accelerate
gradio
huggingface-hub
imageio
imageio-ffmpeg
lpips
matplotlib
packaging
Pillow
pygltflib
rembg[gpu,cli]
rich
safetensors
scikit-image
scikit-learn
scipy
tqdm
transformers
trimesh
kiui >= 0.2.3
xatlas
roma
plyfile
imageio[ffmpeg]
PyMCubes
trimesh
megfile
opencv-python
optimum[onnxruntime-gpu]
rembg[gpu,cli]
httpx[socks]
git+https://github.com/NVlabs/nvdiffrast/